Moves duplication function to RPC
